requirements.txt
django
channels
djangorestframework


